Add files via upload
As an Intern, I have created and completed the task  in 5-7  days approximately on  Jupyter Lab Notebook as per the task of Prodigy InfoTech i.e., DATA SCIENCE 
TASK 2. Task : Perform data cleaning and exploratory data analysis (EDA) on a dataset of your choice, such as the Titanic dataset from Kaggle. Explore the relationships between variables and identify patterns and trends in the data.